Update utils_zcolor.py
An error in the original code meant that the diameter of arrows exported into the Jmol script included only the z-component of the J vector. This means the circular theta-component of the J vector was not included in the size of the arrow, and circular currents calculated from the original script will have looked weaker than they are supposed to. The error only affected the diameter of the arrows exported.